requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.6
dotenv==0.9.9
faiss-cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
hf-xet==1.1.7
httplib2==0.22.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
mail-parser==4.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
